Fix: strip the closing tag for custom HTML tags.
