+++ spec/html_with_custom_markup_examples/blockquote.html
@@ -5,3 +5,6 @@
   <p>This is blockquote text.</p>
 </pre>
 <p>And now I&#8217;m back to normal text!</p>
+<div align="center" style="text-align: center">
+  <p>Center with custom tag</p>
+</div>
