+++ spec/html_with_custom_markup_examples/blockquote_default.html
@@ -5,3 +5,6 @@
   <p>This is blockquote text.</p>
 </blockquote>
 <p>And now I&#8217;m back to normal text!</p>
+<div style="text-align: center">
+  <p>Center with custom tag</p>
+</div>
