+++ spec/html_with_custom_markup_examples/custom_markup_for_html.yml
@@ -3,6 +3,7 @@
   :quote: pre
   :example: blockquote
   :src: blockquote
+  :center: div align="center"
 :Tags:
   "*":
     :open: strong
